ref: Cleanup configs
