+++ vite.main.config.mjs
@@ -2,37 +2,31 @@ import { defineConfig, mergeConfig } from "vite";
 import { external, getBuildConfig, getDefineKeys, pluginHotRestart } from "./vite.base.config.mjs";
 
 /** Vite configuration for the main process */
-export default defineConfig((env) => {
-  /** @type {import('vite').ConfigEnv<'build'>} */
-  const forgeEnv = env;
-  const { forgeConfigSelf } = forgeEnv;
-  const define = getBuildDefine(forgeEnv);
-  return mergeConfig(getBuildConfig(forgeEnv), {
+export default defineConfig((env) =>
+  mergeConfig(getBuildConfig(env), {
     build: {
       lib: {
-        entry: forgeConfigSelf.entry,
+        entry: env.forgeConfigSelf.entry,
         fileName: () => "[name].js",
         formats: ["cjs"],
       },
-      rollupOptions: {
-        external,
-      },
+      rollupOptions: { external },
     },
     plugins: [pluginHotRestart("restart")],
-    define,
+    define: getBuildDefine(env),
     resolve: {
       // Load the Node.js entry.
       mainFields: ["module", "jsnext:main", "jsnext"],
     },
-  });
-});
+  })
+);
 
 /** @type {(env: import('vite').ConfigEnv<'build'>) => Record<string, any>} */
-export const getBuildDefine = (env) => {
-  const { command, forgeConfig } = env;
+export const getBuildDefine = ({ command, forgeConfig }) => {
   const names = forgeConfig.renderer.filter(({ name }) => name != null).map(({ name }) => name);
   const defineKeys = getDefineKeys(names);
-  const define = Object.entries(defineKeys).reduce((acc, [name, keys]) => {
+
+  return Object.entries(defineKeys).reduce((acc, [name, keys]) => {
     const { VITE_DEV_SERVER_URL, VITE_NAME } = keys;
     const def = {
       [VITE_DEV_SERVER_URL]:
@@ -41,6 +35,4 @@ export const getBuildDefine = (env) => {
     };
     return { ...acc, ...def };
   }, {});
-
-  return define;
 };
