+++ vite.preload.config.mjs
@@ -2,20 +2,15 @@ import { defineConfig, mergeConfig } from "vite";
 import { external, getBuildConfig, pluginHotRestart } from "./vite.base.config.mjs";
 
 /** Vite configuration for the preload process */
-// TODO: Can we clean this up at all?
-export default defineConfig((env) => {
-  /** @type {import('vite').ConfigEnv<'build'>} */
-  const forgeEnv = env;
-  const { forgeConfigSelf } = forgeEnv;
-  return mergeConfig(getBuildConfig(forgeEnv), {
+export default defineConfig((env) =>
+  mergeConfig(getBuildConfig(env), {
     build: {
       rollupOptions: {
         external,
         // Preload scripts may contain Web assets, so use the `build.rollupOptions.input` instead `build.lib.entry`.
-        input: forgeConfigSelf.entry,
+        input: env.forgeConfigSelf.entry,
         output: {
-          // TODO: Switch to ESM modules
-          format: "cjs",
+          format: "cjs", // TODO: Switch to ESM modules
           // It should not be split chunks.
           inlineDynamicImports: true,
           entryFileNames: "[name].js",
@@ -25,5 +20,5 @@ export default defineConfig((env) => {
       },
     },
     plugins: [pluginHotRestart("reload")],
-  });
-});
+  })
+);
